Refactor: Collection object for reviews.
Since there is so much work that needs to be done with user
reviews (de duping, sorting, comparing across users), here
is an object that takes these responsibilities.

I still have to integrate this into the main experiment, but it
has the potential to simplify things quite a bit.

This also really simplified the data loading for the small
experiment.